requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
jinja2==3.1.6
openai==1.108.1
pillow==11.3.0
pytesseract==0.3.13
pydantic-settings==2.10.1

